重构File&Directory模块,删除OS_File&OS_Directory实现;重新设计windows平台编译相关projects,支持vs2008,vs2012编译;Log支持logFile配置目录,运行中会自动检测并创建;其它各种杂项更新

+++ llbc/include/llbc/core/log/LogFileAppender.h
@@ -116,14 +116,14 @@ private:
     int _fileBufferSize;
     bool _isDailyRolling;
 
-    size_t _maxFileSize;
+    long _maxFileSize;
     int _maxBackupIndex;
 
 private:
     LLBC_String _fileName;
 
     LLBC_File *_file;
-    size_t _fileSize;
+    long _fileSize;
 
     sint64 _nonFlushLogCount;
     sint64 _logfileLastCheckTime;
